Move drawing primitives to paz.backend.draw; restore __main__ guards
Address review:
- Relocate blend_background, points_in_polygon, fill_polygon and add_occlusion
  from paz.backend.image to paz.backend.draw (drawing/compositing primitives,
  next to overlay_masks); randomize_rendered_image calls them via paz.draw.
  Move their tests to draw_test.py.
- demo.py and train.py: classic 'if __name__ == "__main__":' guard with the
  body inlined (no nested main()).

Verified: draw+image tests pass (20), randomize_rendered_image works through
paz.draw with no import cycle, train smoke loss decreases, retrieval correct.
