+++ paz/backend/image_test.py
@@ -60,40 +60,12 @@ def test_make_random_plain_image_is_uniform_color():
     assert jp.all(image[0, 0] == image[5, 4])
 
 
-def test_blend_background_respects_mask():
-    foreground = jp.full((4, 4, 3), 200, jp.uint8)
-    background = jp.zeros((4, 4, 3), jp.uint8)
-    mask = jp.zeros((4, 4)).at[0, 0].set(1.0)
-    blended = paz.image.blend_background(foreground, background, mask)
-    assert int(blended[0, 0, 0]) == 200
-    assert int(blended[1, 1, 0]) == 0
-
-
-def test_add_occlusion_changes_image_and_keeps_shape():
-    image = jp.full((32, 32, 3), 200, jp.uint8)
-    occluded = paz.image.add_occlusion(jax.random.PRNGKey(1), image)
-    assert occluded.shape == (32, 32, 3)
-    assert not bool(jp.all(occluded == 200))
-
-
 def test_randomize_rendered_image_keeps_shape():
     image = jp.full((32, 32, 3), 128, jp.uint8)
     mask = jp.ones((32, 32))
     key = jax.random.PRNGKey(2)
     output = paz.image.randomize_rendered_image(key, image, mask)
     assert output.shape == (32, 32, 3)
-
-
-def test_fill_polygon_matches_cv2():
-    polygon = jp.array([[12, 10], [50, 16], [55, 48], [28, 58], [10, 38]],
-                       jp.float32)
-    image = jp.full((64, 64, 3), 200, jp.uint8)
-    filled = paz.image.fill_polygon(image, polygon, jp.zeros(3, jp.uint8))
-    reference = np.full((64, 64, 3), 200, np.uint8)
-    cv2.fillPoly(reference, [np.asarray(polygon).astype(np.int32)], (0, 0, 0))
-    jax_mask = np.asarray((filled == 0).all(-1))
-    reference_mask = (reference == 0).all(-1)
-    assert (jax_mask == reference_mask).mean() > 0.95
 
 
 def test_apply_gaussian_blur_matches_cv2_interior():
